HOTFIX: Param store issues (#3)
We were only reading the first 10 results from a param store query, and
not actually using the paged results.  This changes the logic to do the
paged requests, so we don't miss anything.

Also, a method was  added to the static asset lib to allow for
retrieving a specific page.